Support int64 as priority in PQueue

+++ pqueue.go
@@ -15,8 +15,8 @@ const (
 
 type Item struct {
 	value    interface{}
-	priority int
-	index    int
+	priority int64
+	index    uint
 }
 
 // PQueue is a heap priority queue data structure implementation.
@@ -25,11 +25,11 @@ type Item struct {
 type PQueue struct {
 	sync.RWMutex
 	items      []*Item
-	elemsCount int
-	comparator func(int, int) bool
+	elemsCount uint
+	comparator func(int64, int64) bool
 }
 
-func newItem(value interface{}, priority int) *Item {
+func newItem(value interface{}, priority int64) *Item {
 	return &Item{
 		value:    value,
 		priority: priority,
@@ -43,7 +43,7 @@ func (i *Item) String() string {
 // NewPQueue creates a new priority queue with the provided pqtype
 // ordering type
 func NewPQueue(pqType PQType) *PQueue {
-	var cmp func(int, int) bool
+	var cmp func(int64, int64) bool
 
 	if pqType == MAXPQ {
 		cmp = max
@@ -62,7 +62,7 @@ func NewPQueue(pqType PQType) *PQueue {
 }
 
 // Push the value item into the priority queue with provided priority.
-func (pq *PQueue) Push(value interface{}, priority int) *Item {
+func (pq *PQueue) Push(value interface{}, priority int64) *Item {
 	item := newItem(value, priority)
 
 	pq.Lock()
@@ -76,7 +76,7 @@ func (pq *PQueue) Push(value interface{}, priority int) *Item {
 
 // Pop and returns the highest/lowest priority item (depending on whether
 // you're using a MINPQ or MAXPQ) from the priority queue
-func (pq *PQueue) Pop() (interface{}, int) {
+func (pq *PQueue) Pop() (interface{}, int64) {
 	pq.Lock()
 	defer pq.Unlock()
 
@@ -96,7 +96,7 @@ func (pq *PQueue) Pop() (interface{}, int) {
 
 // Head returns the highest/lowest priority item (depending on whether
 // you're using a MINPQ or MAXPQ) from the priority queue
-func (pq *PQueue) Head() (interface{}, int) {
+func (pq *PQueue) Head() (interface{}, int64) {
 	pq.RLock()
 	defer pq.RUnlock()
 
@@ -111,7 +111,7 @@ func (pq *PQueue) Head() (interface{}, int) {
 }
 
 // Size returns the elements present in the priority queue count
-func (pq *PQueue) Size() int {
+func (pq *PQueue) Size() uint {
 	pq.RLock()
 	defer pq.RUnlock()
 	return pq.size()
@@ -141,23 +141,23 @@ func (pq *PQueue) Remove(i *Item) bool {
 	return true
 }
 
-func (pq *PQueue) size() int {
+func (pq *PQueue) size() uint {
 	return pq.elemsCount
 }
 
-func max(i, j int) bool {
+func max(i, j int64) bool {
 	return i < j
 }
 
-func min(i, j int) bool {
+func min(i, j int64) bool {
 	return i > j
 }
 
-func (pq *PQueue) less(i, j int) bool {
+func (pq *PQueue) less(i, j uint) bool {
 	return pq.comparator(pq.items[i].priority, pq.items[j].priority)
 }
 
-func (pq *PQueue) exch(i, j int) {
+func (pq *PQueue) exch(i, j uint) {
 	var tmpItem *Item = pq.items[i]
 
 	pq.items[i] = pq.items[j]
@@ -166,7 +166,7 @@ func (pq *PQueue) exch(i, j int) {
 	pq.items[j].index = j
 }
 
-func (pq *PQueue) swim(k int) {
+func (pq *PQueue) swim(k uint) {
 	for k > 1 && pq.less(k/2, k) {
 		pq.exch(k/2, k)
 		k = k / 2
@@ -174,9 +174,9 @@ func (pq *PQueue) swim(k int) {
 
 }
 
-func (pq *PQueue) sink(k int) {
+func (pq *PQueue) sink(k uint) {
 	for 2*k <= pq.size() {
-		var j int = 2 * k
+		var j uint = 2 * k
 
 		if j < pq.size() && pq.less(j, j+1) {
 			j++
